+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 
 ### Fixed
+- `SearchTrait::List()` may miss some items [#2](https://gitlab.com/wearestancer/library/lib-php/-/issues/2)
 - CI